+++ src/guimain.c
@@ -187,7 +187,11 @@ void guiMainUpdateSideBar() {
     int maxIndex = ctx.sidebarMaxY - 2;
     char text[ctx.sidebarMaxX - 2]; 
     if(ctx.fileList->size <= 0) {
-        mvwprintw(ctx.sidebarPad, 0, 1, "No files found in: \n %s", ctx.fileList->path);
+        if(ctx.fileList->singleFile) {
+            mvwprintw(ctx.sidebarPad, 0, 1, "File not found: \n %s", ctx.fileList->path);
+        } else {
+            mvwprintw(ctx.sidebarPad, 0, 1, "No files found in: \n %s", ctx.fileList->path);
+        }
     } else {
         for(int i = 0 ; i < ctx.fileList->size ; i++) {
             if(ctx.sidebarIdx == i) {